+++ tests/core/test_history.py
@@ -84,7 +84,7 @@ def check_invariants(hist):
 
     # advanced
     assert hist.columns.equals(pd.Index(range(len(hist))))
-    assert isinstance(hist.max(), pd.Series)
+    assert isinstance(hist.squeeze(), pd.Series)
 
 
 def is_equal(hist1: History, hist2: History):
@@ -235,4 +235,4 @@ def test_append_force(__hist, s, max_val):
     hist = __hist
     hist.append(s)
     check_invariants(hist)
-    assert all(hist.max() == max_val)
+    assert all(hist.squeeze() == max_val)